refactor: atomizacion del modulo chat
